Handle coverage failure because of uncaptured inline methods
