+++ orchagent/Makefile.am
@@ -108,6 +108,7 @@ orchagent_SOURCES = \
             nvgreorch.cpp \
             zmqorch.cpp \
             dash/dashenifwdorch.cpp \
+            dash/dashenifwdinfo.cpp \
             dash/dashorch.cpp \
             dash/dashrouteorch.cpp \
             dash/dashvnetorch.cpp \
