+++ tests/mock_tests/mock_orch_test.cpp
@@ -258,6 +258,10 @@ void MockOrchTest::SetUp()
     gDirectory.set(m_VxlanTunnelOrch);
     ut_orch_list.push_back((Orch **)&m_VxlanTunnelOrch);
 
+    m_vnetOrch = new VNetOrch(m_app_db.get(), APP_VNET_TABLE_NAME);
+    gDirectory.set(m_vnetOrch);
+    ut_orch_list.push_back((Orch **)&m_vnetOrch);
+
     ApplyInitialConfigs();
     PostSetUp();
 }
